Build 32-bit (armeabi-v7a) alongside arm64, split apks per ABI
Enable the armeabi-v7a build so the app installs on 32-bit-only devices.
The dynasm JIT only has x86_64/aarch64 backends, so v7a runs the portable
reference interpreter (slower, still correct); arm64 keeps the JIT.

statfs' f_blocks/f_bavail are 32-bit on 32-bit targets and 64-bit on
64-bit ones, so widen them through u64::from to type-check on both. On
the 64-bit build that is an identity conversion, hence the allow.

Split output into one apk per ABI so the native .so - the bulk of the
package - is not duplicated into a single fat apk (which would roughly
double its size once v7a joins arm64). universalApk stays on for
sideloading where the target ABI is not known ahead of time.

+++ wie_android/src/filesystem.rs
@@ -21,6 +21,11 @@ mod host {
 
     /// The block size, the total blocks and the blocks an unprivileged
     /// application may still use, as `statfs` reports them.
+    //
+    // `f_blocks`/`f_bavail` are 32-bit on 32-bit targets (armeabi-v7a) and
+    // 64-bit on 64-bit ones. Widening through `u64::from` type-checks on both;
+    // on the 64-bit build it is an identity conversion, hence the allow.
+    #[allow(clippy::useless_conversion)]
     pub fn storage_blocks(path: &Path) -> Option<(u64, u64, u64)> {
         let path = CString::new(path.as_os_str().as_bytes()).ok()?;
         let mut stats = unsafe { core::mem::zeroed::<libc::statfs>() };
@@ -28,7 +33,7 @@ mod host {
             return None;
         }
 
-        Some((stats.f_bsize as u64, stats.f_blocks, stats.f_bavail))
+        Some((stats.f_bsize as u64, u64::from(stats.f_blocks), u64::from(stats.f_bavail)))
     }
 
     pub fn set_mode(path: &Path, mode: u32) -> std::io::Result<()> {
